# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: La Playita de Don Concho - Restaurant Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }
    post:
      operationId: createProduct
      tags: [products]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateProductRequest" }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
...
  /sales:
    post:
      operationId: createSale
      tags: [sales]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSaleRequest" }
      responses:
        "200":
          description: Sale created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sale" }
    get:
      operationId: listSales
      tags: [sales]
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: Sales
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Sale" }
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
        - name: category
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Expenses
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Expense" }
    post:
      operationId: createExpense
      tags: [expenses]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateExpenseRequest" }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Expense" }
  /expenses/{id}:
    put:
      operationId: updateExpense
      tags: [expenses]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateExpenseRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Expense" }
    delete:
      operationId: deleteExpense
      tags: [expenses]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
...
  /analytics/sales:
    get:
      operationId: getSalesReport
      tags: [analytics]
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: Sales report
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SalesReport" }
  /analytics/sales-trend:
    get:
      operationId: getSalesTrend
      tags: [analytics]
      parameters:
        - name: days
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Sales trend
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TrendPoint" }
  /analytics/products:
    get:
      operationId: getProductPerformance
      tags: [analytics]
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: Product performance
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ProductPerformance" }
  /analytics/waiters:
    get:
      operationId: getWaiterPerformance
      tags: [analytics]
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: Waiter performance
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/WaiterPerformance" }
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer }
  schemas:
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
...
    Product:
      type: object
      required: [id, categoryId, name, price, active]
      properties:
        id: { type: integer }
        categoryId: { type: integer }
        categoryName: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        price: { type: number }
        imageUrl: { type: string, nullable: true }
        active: { type: boolean }
    CreateProductRequest:
      type: object
      required: [categoryId, name, price]
      properties:
        categoryId: { type: integer }
        name: { type: string }
        description: { type: string, nullable: true }
        price: { type: number }
        imageUrl: { type: string, nullable: true }
        active: { type: boolean }
...
    CreateSaleRequest:
      type: object
      required: [orderId, payments]
      properties:
        orderId: { type: integer }
        payments:
          type: array
          items:
            type: object
            required: [method, amount]
            properties:
              method: { type: string, enum: [cash, card, transfer] }
              amount: { type: number }
    Sale:
      type: object
      required: [id, orderId, tableId, total, createdAt]
      properties:
        id: { type: integer }
        orderId: { type: integer }
        tableId: { type: integer }
        tableNumber: { type: integer }
        cashierName: { type: string }
        waiterName: { type: string, nullable: true }
        subtotal: { type: number }
        total: { type: number }
        createdAt: { type: string, format: date-time }
        paymentMethods:
          type: array
          items: { type: string }
...
    SalesReport:
      type: object
      required: [from, to, totalSales, transactionCount, sales]
      properties:
        from: { type: string, format: date }
        to: { type: string, format: date }
        totalSales: { type: number }
        transactionCount: { type: integer }
        averageTicket: { type: number }
        sales:
          type: array
          items: { $ref: "#/components/schemas/Sale" }
    Expense:
      type: object
      required: [id, name, amount, category, expenseDate, paymentMethod]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string, nullable: true }
        amount: { type: number }
        category: { type: string }
        supplier: { type: string, nullable: true }
        expenseDate: { type: string, format: date }
        invoiceNumber: { type: string, nullable: true }
        paymentMethod: { type: string }
        createdByName: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    CreateExpenseRequest:
      type: object
      required: [name, amount, category, expenseDate]
      properties:
        name: { type: string }
        description: { type: string, nullable: true }
        amount: { type: number }
        category: { type: string }
        supplier: { type: string, nullable: true }
        expenseDate: { type: string, format: date }
        invoiceNumber: { type: string, nullable: true }
        paymentMethod: { type: string, enum: [cash, card, transfer] }
...
    TrendPoint:
      type: object
      required: [date, sales, expenses]
      properties:
        date: { type: string }
        sales: { type: number }
        expenses: { type: number }
    ProductPerformance:
      type: object
      required: [productId, productName, quantity, revenue]
      properties:
        productId: { type: integer }
        productName: { type: string }
        quantity: { type: integer }
        revenue: { type: number }
    WaiterPerformance:
      type: object
      required: [waiterId, waiterName, salesCount, total]
      properties:
        waiterId: { type: integer }
        waiterName: { type: string }
        salesCount: { type: integer }
        total: { type: number }
...